Update API
Arguments to freeze:
Using [max] and [heads] incurs an extra check to verify which heads are
successors of max. If we know that we want to copy all heads in the next
upper, then we don't need this extra check.

Recovery flag:
By default set to false; set it to true for the first freeze after
opening a store which crashed during a freeze.